requirements.txt
absl-py==1.3.0
asgiref==3.6.0
astroid==2.15.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==41.0.2
defusedxml==0.7.1
dill==0.3.7
Django==4.1.5
django-allauth==0.54.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf_yasg==1.21.7
exceptiongroup==1.1.1
gunicorn==20.1.0
haversine==2.7.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
joblib==1.2.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.1
oauthlib==3.2.2
ortools==9.5.2237
packaging==23.1
pandas==2.0.0
Pillow==9.4.0
platformdirs==3.9.1
pluggy==1.0.0
protobuf==4.23.2
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.5
pytest==7.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.12.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0
